get_started_google_tools/pyproject.toml
[tool.poetry]
name = "portia_get_started_google_tools"
version = "0.1.0"
description = "Get started with Google tools and Portia AI"
authors = ["Portia AI <hello@portialabs.ai>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
portia-sdk-python = "^0.1.1"
python-dotenv = "^1.0.1"
firecrawl-py = "^1.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"